requirements.txt
fastapi
uvicorn
requests
Beautifulsoup4
chromadb
sentence-transformers
google.generativeai
openai-whisper
pydantic
PyPDF2
docx
pytesseract
openai
# how to become an ai engineer?